+++ Source/ASTextNode2.mm
@@ -14,7 +14,6 @@
 #import <AsyncDisplayKit/ASTextNode.h>  // Definition of ASTextNodeDelegate
 
 #import <tgmath.h>
-#import <deque>
 
 #import <AsyncDisplayKit/_ASDisplayLayer.h>
 #import <AsyncDisplayKit/ASDisplayNode+FrameworkPrivate.h>
@@ -30,17 +29,10 @@
 
 #import <AsyncDisplayKit/CoreGraphics+ASConvenience.h>
 #import <AsyncDisplayKit/ASObjectDescriptionHelpers.h>
+#import <AsyncDisplayKit/ASTextCacheKey.h>
 #import <AsyncDisplayKit/ASTextLayout.h>
+#import <AsyncDisplayKit/ASCache.h>
 #import <AsyncDisplayKit/ASThread.h>
-
-@interface ASTextCacheValue : NSObject {
-  @package
-  ASDN::Mutex _m;
-  std::deque<std::tuple<CGSize, ASTextLayout *>> _layouts;
-}
-@end
-@implementation ASTextCacheValue
-@end
 
 /**
  * If set, we will record all values set to attributedText into an array
@@ -77,6 +69,12 @@ static NSString *ASTextNodeTruncationTokenAttributeName = @"ASTextNodeTruncation
   NSAttributedString *_attributedText;
   NSAttributedString *_composedTruncationText;
   NSArray<NSNumber *> *_pointSizeScaleFactors;
+  
+  // If nil, regenerate.
+  NSAttributedString *_preparedAttributedText;
+  
+  // Just a fast cache. May not be valid but good place to check first.
+  ASTextLayout *_lastUsedLayout;
   
   NSString *_highlightedLinkAttributeName;
   id _highlightedLinkAttributeValue;
@@ -228,18 +226,16 @@ static NSArray *DefaultLinkAttributeNames = @[ NSLinkAttributeName ];
 {
   ASDisplayNodeAssert(constrainedSize.width >= 0, @"Constrained width for text (%f) is too  narrow", constrainedSize.width);
   ASDisplayNodeAssert(constrainedSize.height >= 0, @"Constrained height for text (%f) is too short", constrainedSize.height);
-
+  
   ASLockScopeSelf();
-
-  ASTextContainer *container = [_textContainer copy];
-  NSAttributedString *attributedText = self.attributedText;
-  container.size = constrainedSize;
+  
   [self _ensureTruncationText];
+  ASTextContainer *container = [_textContainer copy];
+  container.size = constrainedSize;
   
-  NSMutableAttributedString *mutableText = [attributedText mutableCopy];
-  [self prepareAttributedString:mutableText];
-  ASTextLayout *layout = [ASTextNode2 compatibleLayoutWithContainer:container text:mutableText];
+  ASTextLayout *layout = [self compatibleLayoutWithContainer:container text:[self l_preparedAttributedText]];
   
+  // Is this necessary?
   [self setNeedsDisplay];
   
   return layout.textBoundingSize;
@@ -281,6 +277,7 @@ static NSArray *DefaultLinkAttributeNames = @[ NSLinkAttributeName ];
   if (!ASCompareAssignCopy(_attributedText, attributedText)) {
     return;
   }
+  _preparedAttributedText = nil;
 
   // Since truncation text matches style of attributedText, invalidate it now.
   [self _invalidateTruncationText];
@@ -323,10 +320,17 @@ static NSArray *DefaultLinkAttributeNames = @[ NSLinkAttributeName ];
   return _textContainer.exclusionPaths;
 }
 
-- (void)prepareAttributedString:(NSMutableAttributedString *)attributedString
+- (NSAttributedString *)l_preparedAttributedText
 {
-  ASLockScopeSelf();
+  if (_preparedAttributedText) {
+    return _preparedAttributedText;
+  }
+  
+  if (!_attributedText) {
+    return [[NSAttributedString alloc] init];
+  }
  
+  NSMutableAttributedString *attributedString = [_attributedText mutableCopy];
   // Apply paragraph style if needed
   [attributedString enumerateAttribute:NSParagraphStyleAttributeName inRange:NSMakeRange(0, attributedString.length) options:kNilOptions usingBlock:^(NSParagraphStyle *style, NSRange range, BOOL * _Nonnull stop) {
     if (style == nil || style.lineBreakMode == _truncationMode) {
@@ -350,6 +354,8 @@ static NSArray *DefaultLinkAttributeNames = @[ NSLinkAttributeName ];
     shadow.shadowBlurRadius = _shadowRadius;
     [attributedString addAttribute:NSShadowAttributeName value:shadow range:NSMakeRange(0, attributedString.length)];
   }
+  _preparedAttributedText = [attributedString copy];
+  return _preparedAttributedText;
 }
 
 #pragma mark - Drawing
@@ -360,107 +366,54 @@ static NSArray *DefaultLinkAttributeNames = @[ NSLinkAttributeName ];
   [self _ensureTruncationText];
   ASTextContainer *copiedContainer = [_textContainer copy];
   copiedContainer.size = self.bounds.size;
-  NSMutableAttributedString *mutableText = [self.attributedText mutableCopy] ?: [[NSMutableAttributedString alloc] init];
-  
-  [self prepareAttributedString:mutableText];
   
   return @{
-    @"container": copiedContainer,
-    @"text": mutableText,
-    @"bgColor": self.backgroundColor ?: [NSNull null]
-  };
+           @"container": copiedContainer,
+           @"text": [self l_preparedAttributedText],
+           @"bgColor": self.backgroundColor ?: [NSNull null],
+           // Pass along self as a hack so that we can update _lastUsedLayout.
+           // We'll be careful
+           @"instance" : self
+           };
 }
 
 /**
  * If it can't find a compatible layout, this method creates one.
  *
- * NOTE: Be careful to copy `text` if needed.
+ * NOTE: The `container` you pass into this should be a copy, as it may be
+ * retained by the cache in the event of a cache miss.
  */
-+ (ASTextLayout *)compatibleLayoutWithContainer:(ASTextContainer *)container
+- (ASTextLayout *)compatibleLayoutWithContainer:(ASTextContainer *)container
                                            text:(NSAttributedString *)text NS_RETURNS_RETAINED
-
 {
-  // Allocate layoutCacheLock on the heap to prevent destruction at app exit (https://github.com/TextureGroup/Texture/issues/136)
-  static ASDN::StaticMutex& layoutCacheLock = *new ASDN::StaticMutex;
-  static NSCache<NSAttributedString *, ASTextCacheValue *> *textLayoutCache;
+  // First quick check – if the last layout we computed is useable.
+  // On a cache miss, this method is going to run long and in some
+  // scenarios it may make sense to run concurrently with different
+  // sizes, so don't hold the lock the whole time.
+  auto lastLayout = ASLockedSelf(_lastUsedLayout);
+  if ([lastLayout isCompatibleWithContainer:container text:text]) {
+    return lastLayout;
+  }
+  
+  // All the keys inside of the cache will have their layout set on them.
+  // So the value kind of doesn't matter – if NSCache had a "set" mode
+  // where you used the -member: method, then we would use that.
+  static ASCache<ASTextCacheKey *, ASTextLayout *> *textLayoutCache;
   static dispatch_once_t onceToken;
   dispatch_once(&onceToken, ^{
-    textLayoutCache = [[NSCache alloc] init];
+    textLayoutCache = [[ASCache alloc] init];
+    textLayoutCache.name = @"org.TextureGroup.Texture.TextNode2LayoutCache";
   });
 
-  layoutCacheLock.lock();
-
-  ASTextCacheValue *cacheValue = [textLayoutCache objectForKey:text];
-  if (cacheValue == nil) {
-    cacheValue = [[ASTextCacheValue alloc] init];
-    [textLayoutCache setObject:cacheValue forKey:[text copy]];
-  }
-
-  // Lock the cache item for the rest of the method. Only after acquiring can we release the NSCache.
-  ASDN::MutexLocker lock(cacheValue->_m);
-  layoutCacheLock.unlock();
-
-  CGRect containerBounds = (CGRect){ .size = container.size };
-  {
-    for (auto &t : cacheValue->_layouts) {
-      CGSize constrainedSize = std::get<0>(t);
-      ASTextLayout *layout = std::get<1>(t);
-
-      CGSize layoutSize = layout.textBoundingSize;
-      // 1. CoreText can return frames that are narrower than the constrained width, for obvious reasons.
-      // 2. CoreText can return frames that are slightly wider than the constrained width, for some reason.
-      //    We have to trust that somehow it's OK to try and draw within our size constraint, despite the return value.
-      // 3. Thus, those two values (constrained width & returned width) form a range, where
-      //    intermediate values in that range will be snapped. Thus, we can use a given layout as long as our
-      //    width is in that range, between the min and max of those two values.
-      CGRect minRect = CGRectMake(0, 0, MIN(layoutSize.width, constrainedSize.width), MIN(layoutSize.height, constrainedSize.height));
-      if (!CGRectContainsRect(containerBounds, minRect)) {
-        continue;
-      }
-      CGRect maxRect = CGRectMake(0, 0, MAX(layoutSize.width, constrainedSize.width), MAX(layoutSize.height, constrainedSize.height));
-      if (!CGRectContainsRect(maxRect, containerBounds)) {
-        continue;
-      }
-      if (!CGSizeEqualToSize(container.size, constrainedSize)) {
-        continue;
-      }
-
-      // Now check container params.
-      ASTextContainer *otherContainer = layout.container;
-      if (!UIEdgeInsetsEqualToEdgeInsets(container.insets, otherContainer.insets)) {
-        continue;
-      }
-      if (!ASObjectIsEqual(container.exclusionPaths, otherContainer.exclusionPaths)) {
-        continue;
-      }
-      if (container.maximumNumberOfRows != otherContainer.maximumNumberOfRows) {
-        continue;
-      }
-      if (container.truncationType != otherContainer.truncationType) {
-        continue;
-      }
-      if (!ASObjectIsEqual(container.truncationToken, otherContainer.truncationToken)) {
-        continue;
-      }
-      // TODO: When we get a cache hit, move this entry to the front (LRU).
-      return layout;
-    }
-  }
-
-  // Cache Miss. Compute the text layout.
-  ASTextLayout *layout = [ASTextLayout layoutWithContainer:container text:text];
-
-  // Store the result in the cache.
-  {
-    // This is a critical section. However we also must hold the lock until this point, in case
-    // another thread requests this cache item while a layout is being calculated, so they don't race.
-    cacheValue->_layouts.push_front(std::make_tuple(container.size, layout));
-    if (cacheValue->_layouts.size() > 3) {
-      cacheValue->_layouts.pop_back();
-    }
-  }
-
-  return layout;
+  
+  ASTextCacheKey *key = [[ASTextCacheKey alloc] initWithContainer:container attributedString:text];
+  
+  ASTextLayout *result = [textLayoutCache objectForKey:key constructedWithBlock:^(ASTextCacheKey *keyParam) {
+    return [keyParam createLayout];
+  }];
+  
+  ASLockedSelf(_lastUsedLayout = result);
+  return result;
 }
 
 + (void)drawRect:(CGRect)bounds withParameters:(NSDictionary *)layoutDict isCancelled:(asdisplaynode_iscancelled_block_t)isCancelledBlock isRasterizing:(BOOL)isRasterizing
@@ -468,7 +421,10 @@ static NSArray *DefaultLinkAttributeNames = @[ NSLinkAttributeName ];
   ASTextContainer *container = layoutDict[@"container"];
   NSAttributedString *text = layoutDict[@"text"];
   UIColor *bgColor = layoutDict[@"bgColor"];
-  ASTextLayout *layout = [self compatibleLayoutWithContainer:container text:text];
+  
+  // We try to limit our use of instance, but we should update
+  // its _lastUsedLayout to get MAXIMUM PERFORMANCE
+  ASTextLayout *layout = [layoutDict[@"instance"] compatibleLayoutWithContainer:container text:text];
   
   if (isCancelledBlock()) {
     return;
@@ -517,7 +473,7 @@ static NSArray *DefaultLinkAttributeNames = @[ NSLinkAttributeName ];
   // See discussion in https://github.com/TextureGroup/Texture/pull/396
   ASTextContainer *containerCopy = [_textContainer copy];
   containerCopy.size = self.calculatedSize;
-  ASTextLayout *layout = [ASTextNode2 compatibleLayoutWithContainer:containerCopy text:_attributedText];
+  ASTextLayout *layout = [self compatibleLayoutWithContainer:containerCopy text:_attributedText];
   NSRange visibleRange = layout.visibleRange;
   NSRange clampedRange = NSIntersectionRange(visibleRange, NSMakeRange(0, _attributedText.length));
 
@@ -531,8 +487,8 @@ static NSArray *DefaultLinkAttributeNames = @[ NSLinkAttributeName ];
   }
 
   NSRange effectiveRange = NSMakeRange(0, 0);
-  for (__strong NSString *attributeName in self.linkAttributeNames) {
-    id value = [self.attributedText attribute:attributeName atIndex:range.start.offset longestEffectiveRange:&effectiveRange inRange:clampedRange];
+  for (__strong NSString *attributeName in _linkAttributeNames) {
+    id value = [_attributedText attribute:attributeName atIndex:range.start.offset longestEffectiveRange:&effectiveRange inRange:clampedRange];
     if (value == nil) {
       // Didn't find any links specified with this attribute.
       continue;
@@ -768,7 +724,7 @@ static NSArray *DefaultLinkAttributeNames = @[ NSLinkAttributeName ];
       // See discussion in https://github.com/TextureGroup/Texture/pull/396
       ASTextContainer *containerCopy = [_textContainer copy];
       containerCopy.size = self.calculatedSize;
-      ASTextLayout *layout = [ASTextNode2 compatibleLayoutWithContainer:containerCopy text:_attributedText];
+      ASTextLayout *layout = [self compatibleLayoutWithContainer:containerCopy text:_attributedText];
       visibleRange = layout.visibleRange;
     }
     NSRange truncationMessageRange = [self _additionalTruncationMessageRangeWithVisibleRange:visibleRange];
@@ -1093,8 +1049,7 @@ static NSAttributedString *DefaultTruncationAttributedString()
 }
 
 /**
- * - cleanses it of core text attributes so TextKit doesn't crash
- * - Adds whole-string attributes so the truncation message matches the styling
+ * Adds whole-string attributes so the truncation message matches the styling
  * of the body text
  */
 - (NSAttributedString *)_locked_prepareTruncationStringForDrawing:(NSAttributedString *)truncationString