+++ Source/Private/TextExperiment/Component/ASTextLayout.h
@@ -102,6 +102,10 @@ extern const CGSize ASTextContainerMaxSize;
 /// This modifier is applied to the lines before the layout is completed,
 /// give you a chance to modify the line position. Default is nil.
 @property (nullable, copy) id<ASTextLinePositionModifier> linePositionModifier;
+
+/// The hash of this container, optionally including the size in the hash.
+- (NSUInteger)hashIncludingSize:(BOOL)includeSize;
+
 @end
 
 
@@ -163,7 +167,7 @@ extern const CGSize ASTextContainerMaxSize;
  @param text The text (if nil, returns nil).
  @return A new layout, or nil when an error occurs.
  */
-+ (nullable ASTextLayout *)layoutWithContainerSize:(CGSize)size text:(NSAttributedString *)text;
++ (nullable ASTextLayout *)layoutWithContainerSize:(CGSize)size text:(NSAttributedString *)text NS_RETURNS_RETAINED;
 
 /**
  Generate a layout with the given container and text.
@@ -172,7 +176,7 @@ extern const CGSize ASTextContainerMaxSize;
  @param text      The text (if nil, returns nil).
  @return A new layout, or nil when an error occurs.
  */
-+ (nullable ASTextLayout *)layoutWithContainer:(ASTextContainer *)container text:(NSAttributedString *)text;
++ (nullable ASTextLayout *)layoutWithContainer:(ASTextContainer *)container text:(NSAttributedString *)text NS_RETURNS_RETAINED;
 
 /**
  Generate a layout with the given container and text.
@@ -183,7 +187,7 @@ extern const CGSize ASTextContainerMaxSize;
  length of the range is 0, it means the length is no limit.
  @return A new layout, or nil when an error occurs.
  */
-+ (nullable ASTextLayout *)layoutWithContainer:(ASTextContainer *)container text:(NSAttributedString *)text range:(NSRange)range;
++ (nullable ASTextLayout *)layoutWithContainer:(ASTextContainer *)container text:(NSAttributedString *)text range:(NSRange)range NS_RETURNS_RETAINED;
 
 /**
  Generate layouts with the given containers and text.
@@ -194,7 +198,7 @@ extern const CGSize ASTextContainerMaxSize;
  or nil when an error occurs.
  */
 + (nullable NSArray<ASTextLayout *> *)layoutWithContainers:(NSArray<ASTextContainer *> *)containers
-                                                      text:(NSAttributedString *)text;
+                                                      text:(NSAttributedString *)text NS_RETURNS_RETAINED;
 
 /**
  Generate layouts with the given containers and text.
@@ -208,7 +212,7 @@ extern const CGSize ASTextContainerMaxSize;
  */
 + (nullable NSArray<ASTextLayout *> *)layoutWithContainers:(NSArray<ASTextContainer *> *)containers
                                                       text:(NSAttributedString *)text
-                                                     range:(NSRange)range;
+                                                     range:(NSRange)range NS_RETURNS_RETAINED;
 
 - (instancetype)init UNAVAILABLE_ATTRIBUTE;
 + (instancetype)new UNAVAILABLE_ATTRIBUTE;
@@ -548,6 +552,12 @@ extern const CGSize ASTextContainerMaxSize;
 - (void)drawInContext:(nullable CGContextRef)context
                  size:(CGSize)size
                 debug:(nullable ASTextDebugOption *)debug;
+
+/**
+ * Whether this layout can be used for the given container-text pair.
+ */
+- (BOOL)isCompatibleWithContainer:(ASTextContainer *)container
+                             text:(NSAttributedString *)text;
 
 @end